updated saveOrUpdate to return saved entity
The `saveOrUpdate` function in `MainRepository` now returns the saved entity, allowing its ID to be accessed.

`SMSParser` is now a `BaseEntity` and managed by its own `SMSParserRepo`.

The `AccountViewModel` has been updated to use `SMSParserRepo` for saving SMS parsers and associates them with an `SMSOperator` by its ID. It now creates an `SMSOperator` if one doesn't exist for the given address.

Signed-off-by: connorT6 <snjp.9702@gmail.com>